+++ src/javascripts/buildRiver.js
@@ -1,5 +1,4 @@
 import printToDom from './printToDom';
-import bearList from './bearList';
 
 const buildRiver = (arr) => {
   let domString = '';
@@ -17,29 +16,4 @@ const buildRiver = (arr) => {
   printToDom('#river', domString);
 };
 
-const formContent = (e) => {
-  e.preventDefault();
-
-  const nameInput = document.querySelector('#bearName').value;
-  const imageInput = document.querySelector('#bearImage').value;
-  const bearIds = bearList
-    .map((bear) => bear.id)
-    .sort((a, b) => a - b);
-  const id = bearIds.length ? bearIds[bearIds.length - 1] + 1 : 1;
-
-  const newBear = {
-    title: nameInput,
-    image: imageInput,
-    id,
-  };
-
-  bearList.push(newBear);
-  buildRiver(bearList);
-  document.querySelector('form').reset();
-};
-
-const buttonEvents = () => {
-  document.querySelector('#submitBtn').addEventListener('click', formContent);
-};
-
-export default buttonEvents;
+export default buildRiver;
