+++ src/javascripts/buttonEvents.js
@@ -1,0 +1,7 @@
+import formContent from './formContent';
+
+const buttonEvents = () => {
+  document.querySelector('#submitBtn').addEventListener('click', formContent);
+};
+
+export default buttonEvents;
