+++ src/javascripts/main.js
@@ -1,6 +1,6 @@
 import '../styles/main.scss';
 import bearForm from './bearForm';
-import buttonEvents from './buildRiver';
+import buttonEvents from './buttonEvents';
 
 const init = () => {
   bearForm();
